Added new and updated audio assets for weapons, ability loops, and idle loops. Updated unit visuals. Cleaned up project assets and project folder structure. Added new UI icons.
Rewrote Turret code to be more consistent, reliable, and extensible in the near future. As an example of this, added an in-game command to manually stop a unit from targeting.
Rewrote construction code to avoid edge case bugs. Construction now costs resources and resources are refunded when built units die.
Rewrote range resist logic to fail less and be more adjustable. Range resist now only absorbs up to a certain ratio of base damage.