Adjusted positioning of the destroyer's engine visuals. Made engine sounds quieter for all units. Removed High Pass Filter components because they are prone to causing bugs. Adjusted micro movement of frigate swarm fighters to be more consistent regardless of framerate and to be more realistic and smooth. Improved Effect_Point to handle a missing AudioSource and also be capable of instantly toggling effects which are reliant on loop prewarming and long particle lifespans. Removed miscellaneous commented-out Debug.Log()s. Adjusted spawning visuals to be more condensed and appealing.
ShieldProject has been implemented in a separate component, more completely implementing what the ability should do. ShieldProject can only be cast at a certain distance and will return the shield if the target leaves a different, larger radius. If a shield is broken, it is returned to the projector. Shield pools can go negative depending on the size of the damage tick that results in their destruction. Infrastructure has been set in place to handle multiple shields on a unit thanks to ShieldMods. Added new visual effect for the target of ShieldProject, which resembles a bubble of energy. This effect orients itself with whatever it is attached to and is sized to fit around any target unit. This effect is mostly transparent to not obstruct gameplay.

Added a custom HPBar addon for the Apollo unit, which displays the current strength of its projected/stored shield. Adjusted "fast update" logic in HPBar and HPBar addons. Updated shield color to be turquoise rather than sky blue to differentiate it from the blue engine colors of some ships.

Created several new audio assets: new engine sound for the corvette; new sound for destroying a projected shield; new sound for shield idle loop; new sound for first projecting a shield; and new sound for returning shield.

Completely repealed legacy AbilityOld system and now entirely relies on Components which extend the new Ability class. Temporarily, ability information in EntityStats is disabled and cooldowns have not yet been re-implemented.