Added the Tooltip window, which provides context-sensitive information about the currently hovered UI element. This includes unit lore, ability descriptions, explanations of game mechanics, and status descriptions.
Changed how units change heights. Now, the player can hold a key to enable a height-change preview mode, which draws lines to the mouse cursor and the height to which the unit will move. Upon releasing the key, the unit will move to that height. For now, this functionality is only available when a single unit is selected. Because grids are no longer used to change heights, only a single grid is needed for horizontal unit movement. This grid moves vertically with the current selection and defaults to a resting position at the initial height that flagships start on.

Tweaked UI backgrounds to be more opaque.