Streamlined box select code. Fixed bug where box selecting with an existing selection would not correctly clear it.
Balance pass: StarLark health reduced to 50 from 100. The ability for the player to easily control groups of units made it easier to execute special maneuvers with StarLarks. Its clear that 6 fast, 200 HP units flying in and pulling away one of your units was too strong, and reducing the health of each StarLark is a step in adding counterplay to that strategy.