Turret Fire() functionality is now handled by its subclasses. The first subclass, Turret_Projectile, behaves exactly as the previous Turret class did, spawning projectiles on every Fire() call. Future subclasses will have unique ways of dealing damage besides firing projectiles.
When a unit dies it passes on a ratio of its current velocity to the wreck it creates. When a unit is killed by superlaser damage it does not create a wreck. Group movement orders within a unit's selection circle now behave correctly and do not turn that unit to a strange angle. Fixed bug where in rare circumstances HP effects were left behind after a unit died. Increased superlaser range to 60 from 25.