Added Objectives. These are Entities which search for units in a radius, and can change ownership depending on which team has more units nearby. Each unit type counts for a different amount of "capture points."
The first objective type grants the capturing team resources, and can remove resources if the previous owner loses control. Resources may go negative if the commander which previously had control over an objective spent all of the resources which were granted and then removed.

Fixed a bug where Controller_Commander would incorrectly check if the selected entity was a unit. Fixed a bug where the damage dealing managers (Manager_Hitscan and Manager_Projectiles) would not check if the impacted collider belonged to a unit or not.